drm/linux: u64 is unsigned long long
* Some basic Linux types were badly defined, causing unnecessary
  differences in various lines of driver code such as kprintf
  statements

* Make u64 and __u64 unsigned long longs like in Linux

* Reduce differences with Linux 4.6 in the generic drm code as well
  as drm/i915 and drm/radeon drivers

* While there, make the include/ directory hierarchy more similar
  to the one in Linux 4.6

+++ sys/dev/agp/agp_i810.c
@@ -1629,8 +1629,8 @@ intel_gtt_clear_range(u_int first_entry, u_int num_entries)
 	agp_intel_gtt_clear_range(intel_agp, first_entry, num_entries);
 }
 
-void intel_gtt_get(size_t *gtt_total, size_t *stolen_size,
-		   phys_addr_t *mappable_base, unsigned long *mappable_end)
+void intel_gtt_get(u64 *gtt_total, size_t *stolen_size,
+		   phys_addr_t *mappable_base, u64 *mappable_end)
 {
 	struct agp_info ainfo;
